+++ src/OpenLoco/src/Ui/Windows/Vehicle.cpp
@@ -4974,6 +4974,14 @@ namespace OpenLoco::Ui::Windows::Vehicle
                 return;
             }
 
+            Vehicles::Vehicle train(*head);
+            EntityId viewportFollowEntity = train.veh2->id;
+            auto main = Ui::WindowManager::getMainWindow();
+            if (main->viewportIsFocusedOnEntity(viewportFollowEntity))
+            {
+                main->viewportUnfocusFromEntity();
+            }
+
             GameCommands::setErrorTitle(StringIds::cant_remove_string_id);
             FormatArguments args{};
             args.skip(10);